Ispis greske prilikom ulaza neautentifikovanog korisnika

+++ src/main/java/ftn/project/controller/DoctorController.java
@@ -28,18 +28,23 @@ public class DoctorController {
 
 	private final UserService userService;
 	private final VacationRequestService vqService;
+	
+	
+	
+	
 
 	@GetMapping("/doctors")
 	public ModelAndView showUsers(HttpServletRequest request,@ModelAttribute("userDto") UserDto userDto, ModelMap model) {
 		
-		String username = (String) request.getSession().getAttribute("logUsername");
-		if(username==null) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
 			return new ModelAndView("badUser");
 		}
 		else {
-		UserDto tempUser =  userService.getUserByUsername(username);
-		
-		if(tempUser.getRoleDto().equals("Clinic Administrator") ) {
+			//autorizacija
+			if(userService.getCA()){
 		
 		model.addAttribute("doctorsDto", userService.allMedicalStaff());
 		return new ModelAndView("doctors", "Model", userService.allMedicalStaff());
@@ -54,23 +59,53 @@ public class DoctorController {
 	
 
 	@PostMapping("/doctors/create")
-	public String createClinic(@Valid @ModelAttribute("userDto") UserDto userDto) {
+	public String createClinic(HttpServletRequest request, @Valid @ModelAttribute("userDto") UserDto userDto) {
+		
+
 		userDto.setRoleDto("doktor");
 		userDto.setPrviLoginDto(false);
 		userService.createUser(userDto);
 		return "redirect:/doctors";
 	}
 
+	
 	@GetMapping("/doctors/delete/{idDto}")
-	public String deleteDoctor(@PathVariable("idDto") Long idDto, ModelMap model) {
+	public String deleteDoctor(HttpServletRequest request,@PathVariable("idDto") Long idDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return "redirect:/badUser";
+		}
+		else {
+			//autorizacija
+			if(userService.getCA()){
+		
 		userService.deleteUser(idDto);
 		return "redirect:/doctors";
+			}else {
+				return "redirect:/badUser";
+			}
+		}
 	}
 	
 	@GetMapping("/doctors/edit/{idDto}")
-	public String getEditPage(@PathVariable("idDto") Long idDto, ModelMap model) {
+	public String getEditPage(HttpServletRequest request,@PathVariable("idDto") Long idDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return "redirect:/badUser";
+		}
+		else {
+		//autorizacija
+		if(userService.getCA()){
 		model.addAttribute("userDto",userService.getUserById(idDto));
 		return "doctorEdit";
+		}else {
+			return "redirect:/badUser";
+		}
+		}
 	}
 	
 	@PostMapping("/doctors/edit/create")
@@ -84,6 +119,15 @@ public class DoctorController {
 	
 	@GetMapping("/doctorsSearch")
 	public ModelAndView searchDoctor(HttpServletRequest request, @ModelAttribute("doctorDto") UserDto doctorDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+		//autorizacija
+		if(userService.getCA()){
 		ModelAndView mav= new ModelAndView("doctorsSearchAll");
 		
 		Set<UserDto> doktori=(Set<UserDto>)request.getSession().getAttribute("doctorsDto");
@@ -94,6 +138,10 @@ public class DoctorController {
 			mav.addObject("doctorsDto", doktori);
 		}
 		return mav;
+		} else {
+			return new ModelAndView("badUser");
+		}
+		}
 	}
 
 	@PostMapping("/doctors/search")
@@ -106,6 +154,16 @@ public class DoctorController {
 	//pretraga pacijenata HomePageDoktor
 	@GetMapping("/patientSearch/doctor")
 	public ModelAndView searchPatientHomePageDoctor(HttpServletRequest request, @ModelAttribute("patientDto") UserDto patientDto, ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+		//autorizacija
+		if(userService.getDoktor()){
+		
 		ModelAndView mav=new ModelAndView("PatientSearchDoctor");
 		
 		Set<UserDto> pacijenti=(Set<UserDto>)request.getSession().getAttribute("patientsDto");
@@ -118,8 +176,13 @@ public class DoctorController {
 		mav.addObject("patientsDto", pacijenti);
 		}
 		return mav;
+		}else {
+			return new ModelAndView("badUser");
+		}
+		}
+		}
 
-	}
+	
 	@PostMapping("/patient/searchD")
 	public String searchPatientR(HttpServletRequest request,@ModelAttribute("patientDto") UserDto patientDto,ModelMap model) {
 		request.getSession().setAttribute("patientsDto", userService.searchPatient(patientDto.nameDto, patientDto.surnameDto, patientDto.getInsuranceNumberDto()));
@@ -131,6 +194,15 @@ public class DoctorController {
 	
 	@GetMapping("/godisnjiOdmorRezervisanje")
 	public ModelAndView rezervisanjeGodisnjeg(HttpServletRequest request,@ModelAttribute("VacationReqDto") VacationRequestDto vacReqDto,ModelMap model) {
+		userService.Autorizacija(request);
+		
+		//autorizacija
+		if(userService.getNull()) {
+			return new ModelAndView("badUser");
+		}
+		else {
+		//autorizacija
+		if(userService.getDoktor()){
 		String username = (String) request.getSession().getAttribute("logUsername");
 		if(username!=null) {
 		UserDto userTemp = userService.getUserByUsername(username);
@@ -142,6 +214,11 @@ public class DoctorController {
 		vacReqDto.setRoleDto("doktor"); }
 	
 		return new ModelAndView("zakazivanjeGodisnjeg");
+		}else {
+			return new ModelAndView("badUser");
+		}
+		
+		}
 
 	}
